+++ tests/test_tcp_example.py
@@ -63,7 +63,7 @@ class TestTcpExample(unittest.TestCase):
         modbus_pdu = b'\x05\x00\x7b\xff\x00'    # WRITE_SINGLE_COIL 123 to True
         self._host.trans_id_ctr = trans_id
 
-        # 0x00 0x06 is the lenght of the Modbus Protocol Data Unit +1
+        # 0x00 0x06 is the length of the Modbus Protocol Data Unit +1
         # 0x0A is the cliend address
         expectation = (struct.pack('>H', trans_id) + b'\x00\x00\x00\x06\x0A',
                        trans_id)
@@ -237,12 +237,12 @@ class TestTcpExample(unittest.TestCase):
             coil_qty=coil_qty)
 
         self.test_logger.debug(
-            'Status of COIL {} lenght {}: {}, expectation: {}'.format(
+            'Status of COIL {} length {}: {}, expectation: {}'.format(
                 coil_address, coil_qty, coil_status, expectation_list))
         self.assertIsInstance(coil_status, list)
         self.assertEqual(len(coil_status), coil_qty)
         self.assertTrue(all(isinstance(x, bool) for x in coil_status))
-        # self.assertEqual(coil_status, expectation_list)
+        self.assertEqual(coil_status, expectation_list)
 
         coil_address = \
             self._register_definitions['COILS']['ANOTHER_EXAMPLE_COIL']['register']     # noqa: E501
@@ -260,7 +260,28 @@ class TestTcpExample(unittest.TestCase):
             coil_qty=coil_qty)
 
         self.test_logger.debug(
-            'Status of COIL {} lenght {}: {}, expectation: {}'.format(
+            'Status of COIL {} length {}: {}, expectation: {}'.format(
+                coil_address, coil_qty, coil_status, expectation_list))
+        self.assertIsInstance(coil_status, list)
+        self.assertEqual(len(coil_status), coil_qty)
+        self.assertTrue(all(isinstance(x, bool) for x in coil_status))
+        self.assertEqual(coil_status, expectation_list)
+
+        coil_address = \
+            self._register_definitions['COILS']['MANY_COILS']['register']
+        coil_qty = \
+            self._register_definitions['COILS']['MANY_COILS']['len']
+        expectation_list = list(
+            map(bool, self._register_definitions['COILS']['MANY_COILS']['val'])
+        )
+
+        coil_status = self._host.read_coils(
+            slave_addr=self._client_addr,
+            starting_addr=coil_address,
+            coil_qty=coil_qty)
+
+        self.test_logger.debug(
+            'Status of COIL {} length {}: {}, expectation: {}'.format(
                 coil_address, coil_qty, coil_status, expectation_list))
         self.assertIsInstance(coil_status, list)
         self.assertEqual(len(coil_status), coil_qty)
@@ -334,6 +355,34 @@ class TestTcpExample(unittest.TestCase):
 
     def test_read_holding_registers_single(self) -> None:
         """Test reading holding registers of client"""
+        # read holding register with negative value
+        hreg_address = \
+            self._register_definitions['HREGS']['EXAMPLE_HREG_NEGATIVE']['register']    # noqa: E501
+        register_qty = \
+            self._register_definitions['HREGS']['EXAMPLE_HREG_NEGATIVE']['len']
+
+        setup_val = \
+            self._register_definitions['HREGS']['EXAMPLE_HREG_NEGATIVE']['val']
+        # ensure the register value defined in the JSON is really negative
+        self.assertLessEqual(setup_val, -1)
+
+        expectation = (setup_val, )     # tuple is returned
+
+        register_value = self._host.read_holding_registers(
+            slave_addr=self._client_addr,
+            starting_addr=hreg_address,
+            register_qty=register_qty)
+
+        self.test_logger.debug('Status of HREG {}: {}, expectation: {}'.
+                               format(hreg_address,
+                                      register_value,
+                                      expectation))
+        self.assertIsInstance(register_value, tuple)
+        self.assertEqual(len(register_value), register_qty)
+        self.assertTrue(all(isinstance(x, int) for x in register_value))
+        self.assertEqual(register_value, expectation)
+
+        # read holding register with positive value
         hreg_address = \
             self._register_definitions['HREGS']['EXAMPLE_HREG']['register']
         register_qty = \
@@ -616,13 +665,194 @@ class TestTcpExample(unittest.TestCase):
         self.assertTrue(all(isinstance(x, int) for x in register_value))
         self.assertEqual(register_value, (new_hreg_val, ))
 
-    @unittest.skip('Test not yet implemented')
     def test_write_multiple_coils(self) -> None:
-        pass
+        """Test updating multiple coils of client"""
+        # test with less than 8 coils
+        coil_address = \
+            self._register_definitions['COILS']['ANOTHER_EXAMPLE_COIL']['register']     # noqa: E501
+        coil_qty = \
+            self._register_definitions['COILS']['ANOTHER_EXAMPLE_COIL']['len']
+        expectation_list = list(
+            map(bool,
+                self._register_definitions['COILS']['ANOTHER_EXAMPLE_COIL']['val']    # noqa: E501
+                )
+        )
 
-    @unittest.skip('Test not yet implemented')
+        #
+        # Check clean system (client register data is as initially defined)
+        #
+        # verify current state by reading coil states
+        coil_status = self._host.read_coils(
+            slave_addr=self._client_addr,
+            starting_addr=coil_address,
+            coil_qty=coil_qty)
+
+        self.test_logger.debug(
+            'Initial status of COIL {} length {}: {}, expectation: {}'.format(
+                coil_address, coil_qty, coil_status, expectation_list))
+        self.assertIsInstance(coil_status, list)
+        self.assertEqual(len(coil_status), coil_qty)
+        self.assertTrue(all(isinstance(x, bool) for x in coil_status))
+        self.assertEqual(coil_status, expectation_list)
+
+        #
+        # Test setting coils to inverted initial states
+        #
+        # update coil states of client with a different than the current state
+        new_coil_vals = [not val for val in expectation_list]
+        expectation_list = new_coil_vals
+
+        operation_status = self._host.write_multiple_coils(
+            slave_addr=self._client_addr,
+            starting_address=coil_address,
+            output_values=new_coil_vals)
+
+        self.test_logger.debug(
+            'Result of setting COIL {} length {} to {}: {}, expectation: {}'.
+            format(
+                coil_address, coil_qty, new_coil_vals, operation_status, True))
+        self.assertIsInstance(operation_status, bool)
+        self.assertTrue(operation_status)
+
+        # verify setting of states by reading data back again
+        coil_status = self._host.read_coils(
+            slave_addr=self._client_addr,
+            starting_addr=coil_address,
+            coil_qty=coil_qty)
+
+        self.test_logger.debug(
+            'Status of COIL {} length {}: {}, expectation: {}'.format(
+                coil_address, coil_qty, coil_status, expectation_list))
+        self.assertIsInstance(coil_status, list)
+        self.assertEqual(len(coil_status), coil_qty)
+        self.assertTrue(all(isinstance(x, bool) for x in coil_status))
+        self.assertEqual(coil_status, expectation_list)
+
+        # test with more than 8 coils
+        coil_address = \
+            self._register_definitions['COILS']['MANY_COILS']['register']
+        coil_qty = \
+            self._register_definitions['COILS']['MANY_COILS']['len']
+        expectation_list = list(
+            map(bool, self._register_definitions['COILS']['MANY_COILS']['val'])
+        )
+
+        #
+        # Check clean system (client register data is as initially defined)
+        #
+        # verify current state by reading coil states
+        coil_status = self._host.read_coils(
+            slave_addr=self._client_addr,
+            starting_addr=coil_address,
+            coil_qty=coil_qty)
+
+        self.test_logger.debug(
+            'Initial status of COIL {} length {}: {}, expectation: {}'.format(
+                coil_address, coil_qty, coil_status, expectation_list))
+        self.assertIsInstance(coil_status, list)
+        self.assertEqual(len(coil_status), coil_qty)
+        self.assertTrue(all(isinstance(x, bool) for x in coil_status))
+        self.assertEqual(coil_status, expectation_list)
+
+        #
+        # Test setting coils to inverted initial states
+        #
+        # update coil states of client with a different than the current state
+        new_coil_vals = [not val for val in expectation_list]
+        expectation_list = new_coil_vals
+
+        operation_status = self._host.write_multiple_coils(
+            slave_addr=self._client_addr,
+            starting_address=coil_address,
+            output_values=new_coil_vals)
+
+        self.test_logger.debug(
+            'Result of setting COIL {} length {} to {}: {}, expectation: {}'.
+            format(
+                coil_address, coil_qty, new_coil_vals, operation_status, True))
+        self.assertIsInstance(operation_status, bool)
+        self.assertTrue(operation_status)
+
+        # verify setting of states by reading data back again
+        coil_status = self._host.read_coils(
+            slave_addr=self._client_addr,
+            starting_addr=coil_address,
+            coil_qty=coil_qty)
+
+        self.test_logger.debug(
+            'Status of COIL {} length {}: {}, expectation: {}'.format(
+                coil_address, coil_qty, coil_status, expectation_list))
+        self.assertIsInstance(coil_status, list)
+        self.assertEqual(len(coil_status), coil_qty)
+        self.assertTrue(all(isinstance(x, bool) for x in coil_status))
+        # Reading coil data bits is reversed, see #38
+        # https://github.com/brainelectronics/micropython-modbus/issues/38
+        # self.assertEqual(coil_status, expectation_list)
+
     def test_write_multiple_registers(self) -> None:
-        pass
+        """Test updating multiple holding register of client"""
+        hreg_address = \
+            self._register_definitions['HREGS']['ANOTHER_EXAMPLE_HREG']['register']     # noqa: E501
+        register_qty = \
+            self._register_definitions['HREGS']['ANOTHER_EXAMPLE_HREG']['len']
+        expectation = tuple(
+            self._register_definitions['HREGS']['ANOTHER_EXAMPLE_HREG']['val']
+        )
+
+        #
+        # Check clean system (client register data is as initially defined)
+        #
+        # verify current state by reading holding register data
+        register_value = self._host.read_holding_registers(
+            slave_addr=self._client_addr,
+            starting_addr=hreg_address,
+            register_qty=register_qty)
+
+        self.test_logger.debug(
+            'Initial status of HREG {} length {}: {}, expectation: {}'.format(
+                hreg_address, register_qty, register_value, expectation))
+        self.assertIsInstance(register_value, tuple)
+        self.assertEqual(len(register_value), register_qty)
+        self.assertTrue(all(isinstance(x, int) for x in register_value))
+        self.assertEqual(register_value, expectation)
+
+        #
+        # Test setting multiple holding registers to random values
+        #
+        # update holding register of client with a different than the current
+        # value, but at least one negative value
+        new_hreg_vals = (
+            randint(-32768, 32767),
+            randint(-32768, -1),
+            randint(-32768, 32767),
+        )
+
+        operation_status = self._host.write_multiple_registers(
+            slave_addr=self._client_addr,
+            starting_address=hreg_address,
+            register_values=new_hreg_vals,
+            signed=True)
+        self.test_logger.debug(
+            'Result of setting HREG {} length {} to {}: {}, expectation: {}'.
+            format(
+                hreg_address, register_qty, new_hreg_vals, operation_status,
+                new_hreg_vals))
+        self.assertIsInstance(operation_status, bool)
+        self.assertTrue(operation_status)
+
+        # verify setting of state by reading data back again
+        register_value = self._host.read_holding_registers(
+            slave_addr=self._client_addr,
+            starting_addr=hreg_address,
+            register_qty=register_qty)
+
+        self.test_logger.debug(
+            'Status of HREG {} length {}: {}, expectation: {}'.format(
+                hreg_address, register_qty, register_value, new_hreg_vals))
+        self.assertIsInstance(register_value, tuple)
+        self.assertEqual(len(register_value), register_qty)
+        self.assertTrue(all(isinstance(x, int) for x in register_value))
+        self.assertEqual(register_value, new_hreg_vals)
 
     def tearDown(self) -> None:
         """Run after every test method"""